feature[test]: test for issue #25 are moved inside VerifyAtLeastAtMostExactlyTest.kt; tests for issue #35 are moved to a Maps dedicated test class.
