+++ mockk/common/src/test/kotlin/io/mockk/it/VerifyAtLeastAtMostExactlyTest.kt
@@ -8,6 +8,7 @@ import kotlin.test.assertFailsWith
 class VerifyAtLeastAtMostExactlyTest {
     class MockCls {
         fun op(a: Int) = a + 1
+        fun op2(a: Int, b: Int) = a + b
     }
 
     val mock = mockk<MockCls>()
@@ -75,6 +76,18 @@ class VerifyAtLeastAtMostExactlyTest {
         }
     }
 
+    /**
+     * See issue #25
+     */
+    @Test
+    fun exactlyZeroWithAny() {
+        doCalls2()
+
+        verify(exactly = 0) {
+            mock.op2(3, any())
+            mock.op(3)
+        }
+    }
 
     @Test
     fun exactlyOnce() {
@@ -157,6 +170,22 @@ class VerifyAtLeastAtMostExactlyTest {
         assertEquals(2, mock.op(1))
         assertEquals(3, mock.op(1))
         assertEquals(3, mock.op(1))
+    }
+
+    fun doCalls2() {
+        every { mock.op(0) } throws RuntimeException("test")
+        every { mock.op(1) } returnsMany listOf(1, 2, 3) andThen 5
+        every { mock.op2(2, 1) } returns 3
+
+        assertFailsWith(RuntimeException::class) {
+            mock.op(0)
+        }
+
+        assertEquals(1, mock.op(1))
+        assertEquals(2, mock.op(1))
+        assertEquals(3, mock.op(1))
+        assertEquals(5, mock.op(1))
+        assertEquals(3, mock.op2(2, 1))
     }
 
 }
